requirements.txt
google-generativeai
google-cloud-bigquery
fastapi